requirements.txt
tensorflow==2.9.1
numpy==1.23.5
scikit-learn==1.0.2
Pillow==9.3.0
Flask==2.2.2
gunicorn==20.1.0
requests==2.28.1
flask-cors==4.0.0